ファイルの構造を整理
