+++ programs/html/index.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>title</title>
-    <link rel="stylesheet" href="style-title.css">
+    <link rel="stylesheet" href="../css/style-title.css">
 </head>
 <body>
     <h1 id="title">title</h1>
@@ -23,5 +23,5 @@
         「×」ボタンを押すと、ヘルプが閉じます<br>
     </p>
 </body>
-    <script src="title.js"></script>
+    <script src="../program/title/title.js"></script>
 </html>
