+++ programs/html/mondai.html
@@ -3,7 +3,7 @@
 <head>
     <meta charset="UTF-8">
     <title>question&scoring</title>
-    <link rel="stylesheet" href="style-mondai.css">
+    <link rel="stylesheet" href="../css/style-mondai.css">
 </head>
 <body>
     <button id="goEditor" onclick="location.href='editor.html'">エディターへ</button>
@@ -26,8 +26,14 @@
 
     <script src="https://cdnjs.cloudflare.com/ajax/libs/ace/1.4.12/ace.js"></script>
     <script src="https://cdn.jsdelivr.net/pyodide/v0.21.3/full/pyodide.js"></script>
-    <script src="ace.js"></script>
-    <script src="mondai.js"></script>
-    <script src="scoring.js"></script>
+    <script src="../program/base/ace.js"></script>
+    <script src="../program/mondai/mondai.js"></script>
+    <script src="../program/mondai/createMondaiText.js"></script>
+    <script src="../program/mondai/createAnswerText.js"></script>
+    <script src="../program/mondai/scoring.js"></script>
+    <script src="../program/mondai/createAnswerExample.js"></script>
+    <script src="../program/mondai/selectMondai.js"></script>
+    <script src="../program/mondai/getResultOfExe.js"></script>
+    <script src="../program/mondai/addElement.js"></script>
 </body>
 </html>
