+++ programs/program/editor/main.js
@@ -1,14 +1,14 @@
-let pushRun_convert = document.getElementById("run");
-let pushRun_step = document.getElementById("run_step");
-let pushRun_prev = document.getElementById("prev_step");
-let pushRun_next = document.getElementById("next_step");
-let pushRun_loop = document.getElementById("step_loop");
-let pushRun_loop_prev = document.getElementById("prev_loop");
-let pushRun_loop_next = document.getElementById("next_loop");
-let pushShow_execute = document.getElementById("show_execute");
-let pushShow_step = document.getElementById("show_step");
-let pushShow_loop = document.getElementById("show_loop");
-let pushHelp = document.getElementById("help_button");
+const pushRun_convert = document.getElementById("run");
+const pushRun_step = document.getElementById("run_step");
+const pushRun_prev = document.getElementById("prev_step");
+const pushRun_next = document.getElementById("next_step");
+const pushRun_loop = document.getElementById("step_loop");
+const pushRun_loop_prev = document.getElementById("prev_loop");
+const pushRun_loop_next = document.getElementById("next_loop");
+const pushShow_execute = document.getElementById("show_execute");
+const pushShow_step = document.getElementById("show_step");
+const pushShow_loop = document.getElementById("show_loop");
+const pushHelp = document.getElementById("help_button");
 
 let currentStep = 0;
 let codeLines = [];
