+++ programs/program/editor/variables.js
@@ -1,7 +1,8 @@
+// 変数表の更新を行う関数
 let variables = {};
 let variables_step = {};
 let variables_loop = {};
-
+// 一括実行用
 function updateVariableTable(code) {
     const tableBody = document.querySelector('#variables_table tbody');
     tableBody.innerHTML = '<tr><td>変数名</td><td>値</td></tr><tbody></tbody>';
@@ -19,7 +20,7 @@ function updateVariableTable(code) {
         }
     }
 }
-
+// ステップ実行用
 function updateVariableTable_step(code) {
     const tableBody = document.querySelector('#variables_table_output tbody');
     tableBody.innerHTML = '<tr><td>変数名</td><td>値</td></tr><tbody></tbody>';
@@ -37,13 +38,12 @@ function updateVariableTable_step(code) {
         }
     }
 }
-
-
+// ループ実行用
 function updateVariableTable_loop(code){
     const tableBody = document.querySelector('#variables_table_loop tbody');
     tableBody.innerHTML = '<tr><td>変数名</td><td>値</td></tr><tbody></tbody>';
     for (const [name, value] of Object.entries(variables_loop)) {
-        if(code.includes(name) && code.includes('def') === false){
+        if(code.includes(name) && typeof value !== 'function'){
             const row = document.createElement('tr');
             const nameCell = document.createElement('td');
             const valueCell = document.createElement('td');
