--- programs/styles.css
@@ -1,367 +1,0 @@
-/* 全体をリセット */
-body {
-    margin: 0;
-    padding: 0;
-    display: flex;
-    flex-direction: column;
-    height: 100vh;
-    font-family: Arial, sans-serif;
-}
-
-/* エディタ部分の配置 */
-#editor {
-    position: absolute;
-    bottom: 20%;
-    left: 0;
-    width: 50%;
-    height: 50%;
-    border: 2px solid #ccc;
-}
-
-#compile {
-    border-radius: 100% 100% 0 0;
-    position: absolute;
-    bottom: 15.3%;
-    left: 0;
-    width: 12%;
-    height: 5%;
-    border: 2px solid #080cf1;
-    background-color: #080cf1;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#run {
-    border-radius: 100% 100% 0 0;
-    position: absolute;
-    top: 30%;
-    right: 41%;
-    width: 8%;
-    height: 5%;
-    border: 2px solid #08f070;
-    background-color: #08f070;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#reset {
-    border-radius: 0 0 100% 100%;
-    position: absolute;
-    bottom: 15.3%;
-    left: 12%;
-    width: 12%;
-    height: 5%;
-    border: 2px solid #0a0a0a;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-/* 構文解析ログの配置 */
-#result {
-    position: absolute;
-    bottom: 0;
-    left: 0;
-    width: 50%;
-    height: 15%;
-    border: 2px solid #080cf1;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-
-#output {
-    position: absolute;
-    bottom: 35%;
-    right: 0;
-    width: 48%;
-    height: 30%;
-    border: 4px solid #f04a08;
-    border-bottom: none;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#variable_table_output {
-    position: absolute;
-    bottom: 5%;
-    right: 0%;
-    width: 48%;
-    height: 30%;
-    border: 4px solid #f04a08;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#run_step {
-    border-radius: 100% 100% 0 0;
-    position: absolute;
-    top: 30%;
-    right: 41%;
-    width: 8%;
-    height: 5%;
-    border: 2px solid #f04a08;
-    background-color: #f04a08;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#prev_step {
-    border-radius: 100% 100% 0 0;
-    position: absolute;
-    top: 30%;
-    right: 33%;
-    width: 8%;
-    height: 5%;
-    border: 2px solid #f04a08;
-    background-color: #f04a08;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#next_step {
-    border-radius: 100% 100% 0 0;
-    position: absolute;
-    top: 30%;
-    right: 25%;
-    width: 8%;
-    height: 5%;
-    border: 2px solid #f04a08;
-    background-color: #f04a08;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#loop {
-    position: absolute;
-    bottom: 35%;
-    right: 0;
-    width: 48%;
-    height: 30%;
-    border: 4px solid #ecf008;
-    border-bottom: none;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#variable_table_loop {
-    position: absolute;
-    bottom: 5%;
-    right: 0%;
-    width: 48%;
-    height: 30%;
-    border: 4px solid #ecf008;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#step_loop {
-    border-radius: 100% 100% 0 0;
-    position: absolute;
-    top: 30%;
-    right: 41%;
-    width: 8%;
-    height: 5%;
-    border: 2px solid #ecf008;
-    background-color: #ecf008;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#prev_loop {
-    border-radius: 100% 100% 0 0;
-    position: absolute;
-    top: 30%;
-    right: 33%;
-    width: 8%;
-    height: 5%;
-    border: 2px solid #ecf008;
-    background-color: #ecf008;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#next_loop {
-    border-radius: 100% 100% 0 0;
-    position: absolute;
-    top: 30%;
-    right: 25%;
-    width: 8%;
-    height: 5%;
-    border: 2px solid #ecf008;
-    background-color: #ecf008;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#execute {
-    position: absolute;
-    bottom : 35%;
-    right: 0;
-    width: 48%;
-    height: 30%;
-    border: 4px solid #08f070;
-    border-bottom: none;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#variable_table {
-    position: absolute;
-    bottom : 5%;
-    right: 0%;
-    width: 48%;
-    height: 30%;
-    border: 4px solid #08f070;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#variables_table {
-    width: 100%;
-    border-collapse: collapse;
-}
-
-#variables_table th, #variables_table td, #variables_table tbody th, #variables_table tbody td {
-    padding: 6px;
-    text-align: left;
-    border: 4px solid #ddd;
-}
-
-#variables_table_output {
-    width: 100%;
-    border-collapse: collapse;
-}
-
-#variables_table_output th, #variables_table_output td, #variables_table_output tbody th, #variables_table_output tbody td {
-    padding: 6px;
-    text-align: left;
-    border: 4px solid #ddd;
-}
-
-#variables_table_loop {
-    width: 100%;
-    border-collapse: collapse;
-}
-
-#variables_table_loop th, #variables_table_loop td, #variables_table_loop tbody th, #variables_table_loop tbody td {
-    padding: 6px;
-    text-align: left;
-    border: 4px solid #ddd;
-}
-
-.highlight-line {
-    background-color: rgb(255, 81, 0) !important;
-}
-
-.hidden {
-    display: none;
-}
-
-#show_execute {
-    border-radius: 0 0 100% 100%;
-    position: absolute;
-    bottom: 0;
-    right: 32%;
-    width: 16%;
-    height: 5%;
-    border: 4px solid #08f070;
-    background-color: #08f070;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#show_step {
-    border-radius: 0 0 100% 100%;
-    position: absolute;
-    bottom: 0;
-    right: 16%;
-    width: 16%;
-    height: 5%;
-    border: 4px solid #f04a08;
-    background-color: #f04a08;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#show_loop {
-    border-radius: 0 0 100% 100%;
-    position: absolute;
-    bottom: 0;
-    right: 0;
-    width: 16%;
-    height: 5%;
-    border: 4px solid #ecf008;
-    background-color: #ecf008;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#mondai {
-    position: absolute;
-    top: 0;
-    left: 0;
-    width: 98%;
-    height: 28%;
-    border: 4px solid #0e0d0e;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#help_button {
-    /*丸くしたい場合*/
-    border-radius: 50%;
-    position: absolute;
-    top: 29.5%;
-    right: 0;
-    width: 5%;
-    height: 5%;
-    border: 4px solid #0e0d0e;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-#help_text {
-    position: absolute;
-    top: 0;
-    right: 5%;
-    width: 60%;
-    height: 75%;
-    border: 4px solid #0e0d0e;
-    overflow-y: auto;
-    font-size: 18px;
-    display: none;
-    background-color: #ddd;
-}
-
-#convert_button {
-    border-radius: 0 0 100% 100%;
-    position: absolute;
-    bottom: 15.3%;
-    left: 38%;
-    width: 12%;
-    height: 5%;
-    border: 2px solid #0a0a0a;
-    overflow-y: auto;
-    font-size: 18px;
-}
-
-.toggle-btn {
-    color: blue;
-    cursor: pointer;
-    text-decoration: underline;
-}
-
-.problem-text {
-    margin-left: 10px;
-}
-
-.toggle-btn_s{
-    color: blue;
-    cursor: pointer;
-    text-decoration: underline;
-}
-
-.answer-text {
-    margin-left: 10px;
-    padding-left: 20px;
-}
