--- renew/test.txt
@@ -1,3 +1,0 @@
-Alice
-Bob
-Charlie
